Remove -1 default value from Game initial ID
Fixed an issue where you would get a 500 Internal Server Error when sending a POST request due to Cannot insert explicit value for identity column in table 'Games' when IDENTITY_INSERT is set to OFF.

Default Value Conventions: By default, EF Core treats an integer property without any initialization as having a default value of 0. This means that if you don't assign a value to Id in your C# code, EF Core will assume it's 0, and SQL Server will generate a new identity value accordingly.

So, when you set Id to 0, EF Core assumes it's a valid value for a newly inserted record, and SQL Server generates a new identity value as expected. However, when you set Id to -1, EF Core treats it as an explicit value assignment, which conflicts with the identity column's behavior, and SQL Server raises an error.